# src/main/java/com/ex/bookmanagement/dto/CreateCategoryRequest.java
package com.ex.bookmanagement.dto;

import io.swagger.v3.oas.annotations.media.Schema;
import jakarta.validation.constraints.NotBlank;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;

@Getter
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class CreateCategoryRequest {
    @NotBlank(message = "카테고리 이름은 필수입니다.")
    @Schema(example = "문학")
    private String name;
}